Don't load gems in the gem loader when `require` is set to false.
This is probably not the ideal way to fix this. I think the correct
solution would be to have a way for the user to configure certain gems
to not be loaded by Sorbet, because there are totally valid reasons to
want Sorbet to load and scaffold a given gem without it being required
by default by Bundler.